+++ client/src/components/table/ExpirationData/ExpirationDataTable.tsx
@@ -1,4 +1,4 @@
-import React from "react";
+import React, { useState } from "react";
 import {
   Table,
   TableBody,
@@ -7,61 +7,100 @@ import {
   TableRow,
   TableCell,
 } from "@../../components/ui/table";
-import ButtonComponent from "../../CustomButton";
+import ButtonComponent from "src/components/CustomButton";
 import { ExpirationDateHook } from "src/hooks/ExpirationDateHook";
+import { ConfirmDeleteModal } from "src/components/modal/ExpirationDateModal";
+
+/**
+ * @jojayeon 24.08.07
+ * @returns 유통기한 관리 테이블
+ */
 
 export const ExpirationDataTable: React.FC = () => {
-  const { data, loading, error, deleteProduct } = ExpirationDateHook(); //구조 분해 할당
+  const { data, loading, error, deleteProduct } = ExpirationDateHook();
+  const [open, setOpen] = useState(false);
+  const [selectedProductId, setSelectedProductId] = useState<string | null>(
+    null,
+  );
+
   if (loading) {
     return <div>Loading...</div>;
   }
   if (error) {
     return <div>{error}</div>;
   }
-
+  //정렬하기 날짜별, 이름별 2번
   const sortedProducts = data.sort((a, b) => {
-    return (
-      new Date(a.expirationDate).getTime() -
-      new Date(b.expirationDate).getTime()
-    ); //오름차순 으로 정렬한 것
+    const expirationDateComparison = new Date(a.expirationDate).getTime() - new Date(b.expirationDate).getTime();
+
+    if (expirationDateComparison !== 0) {
+      return expirationDateComparison;
+    }
+
+    return a.productName.localeCompare(b.productName);
   });
 
-  const handleDelete = (_id: string) => {
-    deleteProduct(_id.toString());
+  // 도달창 띄우고 취소 확인
+  //확인
+  const handleDelete = () => {
+    if (selectedProductId) {
+      deleteProduct(selectedProductId);
+      setOpen(false);
+    }
+  };
+
+  //폐기 버튼
+  const openModal = (product: string) => {
+    setSelectedProductId(product);
+    setOpen(true);
+  };
+
+  //취소 버튼
+  const closeModal = () => {
+    setOpen(false);
+    setSelectedProductId(null);
   };
 
   return (
-    <Table>
-      <TableHeader>
-        <TableRow>
-          <TableHead>재고명</TableHead>
-          <TableHead className="pr-10%">수량</TableHead>
-          <TableHead className="">유통기한</TableHead>
-          <TableHead className="pr-10 text-right">폐기</TableHead>
-        </TableRow>
-      </TableHeader>
-      <TableBody>
-        {sortedProducts.map((product) => (
-          <TableRow key={product._id.toString()}>
-            <TableCell>{product.productCategory}</TableCell>
-            <TableCell>{product.productName}</TableCell>
-            <TableCell>{product.quantity}</TableCell>
-            <TableCell>
-              {new Date(product.expirationDate).toDateString()}
-            </TableCell>
-            <TableCell className="text-right">
-              <ButtonComponent
-                variant="default"
-                type="button"
-                onClick={() => handleDelete(product._id.toString())}
-              >
-                폐기하기
-              </ButtonComponent>
-              {/* handleDelete 데이터 가져오기*/}
-            </TableCell>
+    <>
+      <Table>
+        <TableHeader>
+          <TableRow>
+            <TableHead>분류</TableHead>
+            <TableHead>제품명</TableHead>
+            <TableHead>수량</TableHead>
+            <TableHead className="">유통기한</TableHead>
+            <TableHead className="text-right pr-10">폐기</TableHead>
           </TableRow>
-        ))}
-      </TableBody>
-    </Table>
+        </TableHeader>
+        <TableBody>
+          {sortedProducts.map((product) => (
+            <TableRow key={product._id}>
+              <TableCell>{product.productCategory}</TableCell>
+              <TableCell>{product.productName}</TableCell>
+              <TableCell>{product.quantity}</TableCell>
+              <TableCell>
+                {new Date(product.expirationDate).toDateString()}
+              </TableCell>
+              <TableCell className="text-right">
+                <ButtonComponent
+                  variant="default"
+                  type="button"
+                  onClick={() => openModal(product._id)}
+                >
+                  폐기하기
+                </ButtonComponent>
+              </TableCell>
+            </TableRow>
+          ))}
+        </TableBody>
+      </Table>
+      {/* 모달창 */}
+      <ConfirmDeleteModal
+        open={open}
+        onClose={closeModal}
+        onConfirm={handleDelete}
+      />
+    </>
   );
 };
