+++ client/src/hooks/ExpirationDateHook.tsx
@@ -23,7 +23,7 @@ export const ExpirationDateHook = () => {
   const fetchData = async () => {
     setLoading(true);
     try {
-      const response = await fetch(url3001Generator(EP_PRODUCTS)); //endpoint - products
+      const response = await fetch(url3001Generator(EP_PRODUCTS));
       if (!response.ok) {
         throw new Error("네트워크 응답이 올바르지 않습니다.");
       }
@@ -44,8 +44,7 @@ export const ExpirationDateHook = () => {
     const deleteUrl = urlJoin(url3001Generator(EP_PRODUCTS, _id));
     try {
       await fetch(deleteUrl, {
-        //엔드 포인트 products/${productID}` - 받는 부분은 어떻게 설정하지 변경해야하나?
-        method: "DELETE", // 이거 때문에 문제가 없는 것 처럼 이야기 하는데 맞는지 모르겠음
+        method: "DELETE", 
       });
       fetchData();
     } catch (err) {
