+++ client/src/pages/admin/stockDate.tsx
@@ -1,9 +1,8 @@
 import React from "react";
 import AdminNav from "src/components/admin/adminNav";
-import { ExpirationDataTable } from "src/components/table/ExpirationDataTable";
+import { ExpirationDataTable } from "src/components/table/ExpirationData/ExpirationDataTable";
 import Logo from "src/components/logo";
 import TitleComponent from "src/components/titleComponent";
-import Modal from "src/components/modal/Modal";
 /**
  * @jojayeon 24.08.05
  * @returns 유통관리 페이지
